+++ arm/src/Hook.h
@@ -14,14 +14,14 @@ class Hook
 
 public:
 	virtual bool evalue() = 0;
-	Hook(uint16_t id, bool isUnique, uint8_t nbCallback);
+	Hook(uint8_t id, bool isUnique, uint8_t nbCallback);
 	virtual ~Hook();
 	void insert(Executable* f, uint8_t indice); // n'est pas surchargé, donc pas besoin de la mettre en virtuel
 	bool execute(); // idem
 	uint16_t getId(); // idem
 
 private:
-	uint16_t m_id;
+	uint8_t m_id;
 
 protected:
 	bool m_isUnique;
@@ -39,7 +39,7 @@ private:
 
 public:
 	static void setDateDebutMatch();
-	HookTemps(uint16_t id, uint8_t nbCallback, uint32_t dateExecution);
+	HookTemps(uint8_t id, uint8_t nbCallback, uint32_t dateExecution);
 	bool evalue();
 };
 
@@ -49,32 +49,32 @@ private:
 	uint8_t m_nbCapteur;
 
 public:
-	HookContact(uint16_t id, bool isUnique, uint8_t nbCallback, uint8_t nbCapteur);
+	HookContact(uint8_t id, bool isUnique, uint8_t nbCallback, uint8_t nbCapteur);
 	bool evalue();
 };
 
 class HookPosition : public Hook
 {
 private:
-	uint32_t m_x;
-	uint32_t m_y;
-	uint32_t m_tolerance;
+	uint16_t m_x;
+	uint16_t m_y;
+	uint16_t m_tolerance;
 
 public:
-	HookPosition(uint16_t id, uint8_t nbCallback, uint32_t x, uint32_t y, uint32_t tolerance);
+	HookPosition(uint8_t id, uint8_t nbCallback, uint16_t x, uint16_t y, uint16_t tolerance);
 	bool evalue();
 };
 
 class HookDemiPlan : public Hook
 {
 private:
-	uint32_t m_x;
-	uint32_t m_y;
-	uint32_t m_direction_x;
-	uint32_t m_direction_y;
+	uint16_t m_x;
+	uint16_t m_y;
+	uint16_t m_direction_x;
+	uint16_t m_direction_y;
 
 public:
-	HookDemiPlan(uint16_t id, uint8_t nbCallback, uint32_t x, uint32_t y, uint32_t direction_x, uint32_t direction_y);
+	HookDemiPlan(uint8_t id, uint8_t nbCallback, uint16_t x, uint16_t y, uint16_t direction_x, uint16_t direction_y);
 	bool evalue();
 };
